requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
isodate==0.6.1
lxml==4.9.3
platformdirs==3.9.1
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.0.4
zeep==4.2.1
